+++ Bullet_Hell/src/ResourceCache/DefaultResourceLoader.h
@@ -15,10 +15,10 @@
 class DefaultResourceLoader : public ResourceLoader
 {
 	virtual bool discard_raw_buffer_after_load();
-	virtual unsigned int get_loaded_resource_size(char* raw_buffer,
-		unsigned int raw_size);
+	virtual size_t get_loaded_resource_size(char* raw_buffer,
+		size_t raw_size);
 	virtual std::string get_pattern();
-	virtual bool load_resource(char* raw_buffer, unsigned int raw_size,
+	virtual bool load_resource(char* raw_buffer, size_t raw_size,
 		std::shared_ptr<ResourceHandle> handle);
 	virtual bool use_raw_file();
 };
