+++ Bullet_Hell/src/ResourceCache/ResourceZipFile.h
@@ -74,5 +74,5 @@ public:
 	/// <returns>The name of the resource at the supplied index.
 	/// Empty if index is not in bounds.
 	/// </returns>
-	virtual std::string get_resource_name(int index);
+	virtual std::string get_resource_name(size_t index);
 };
